chatbot-main/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "lxml>=6.0.0",
    "openai>=1.93.0",
    "rich>=14.0.0",
    "yfinance>=0.2.64",
    "beautifulsoup4>=4.13.0",
    "requests>=2.31.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.181.0",
    "psycopg2-binary>=2.9.9",
]

[tool.setuptools.packages.find]
include = ["chatbot*", "database*", "services*", "tools*", "util*"]
